Create logic for updating queue thumbnails
Also bugfix for reset button not committing changes to database

+++ filmulator-gui/core/imagePipeline.h
@@ -18,8 +18,8 @@ struct ProcessingParameters {
     //demosaic is now valid
 
     std::vector<float> exposureComp;
-    double temperature;
-    double tint;
+    float temperature;
+    float tint;
     //prefilmulation stuff is now valid
 
     filmulateParams filmParams;
